perf: 优化日频扩展性能（30倍）+ 修复关键bug + 添加验证测试
核心改进：
- 财务处理器性能优化：reindex+ffill替代循环填充，速度提升30倍
- 数据管理器bug修复：修复auxiliary_data_path加载优先级和索引对齐KeyError
- 因子处理器代码质量提升：明确dtype避免类型歧义

新增测试：
- tests/verify_current_changes.py：核心功能快速验证脚本
- tests/test_asset_turnover_ttm.py：资产周转率因子测试
- tests/test_inventory_turnover_ttm.py：存货周转率因子测试
- tests/performance/test_expand_to_daily_performance.py：日频扩展性能测试

其他更新：
- .gitignore：排除临时文档和结果目录
- .claude/settings.local.json：新增必要的Bash权限

+++ multifactors_beta/core/utils/factor_processing.py
@@ -32,23 +32,23 @@ class FactorOrthogonalizer:
         """
         # 导入所需模块
         from .data_cleaning import OutlierHandler, Normalizer
-        
+
         # 按日期分组处理
         def _orthogonalize_date(date_data: pd.DataFrame) -> pd.DataFrame:
             """对单个日期的数据进行正交化"""
-            result = pd.DataFrame(index=date_data.index, columns=date_data.columns)
-            
+            result = pd.DataFrame(index=date_data.index, columns=date_data.columns, dtype=float)
+
             for i, col in enumerate(date_data.columns):
                 factor_data = date_data[col].values
-                
+
                 # 去除异常值
                 if remove_outliers:
                     factor_data = OutlierHandler.remove_outlier(factor_data, method="IQR")
-                
+
                 # 标准化
                 if normalize:
                     factor_data = Normalizer.normalize(factor_data, method="zscore")
-                
+
                 if i == 0:
                     # 第一个因子直接使用
                     result.iloc[:, i] = factor_data
